Added second STM code for Receiving signal
The main code was revised to ensure that it could properly send signals depending on the state of the light sensor. The receiver STM32 was programmed to constantly read from the Sender STM32 pin to check if it was an appropriate time to turn on the buzzer. There is also a timer that ensures that turns off the speaker after 15 seconds.